Started refactoring of existing code. Exchange factory pattern
Added coinbase and kraken repositories as submodules.
Added a script to install exchange wrappers locally to the project (so, Thirdparty/install.sh needs to be run always that any exchange package is updated).
Added a README file, with the instructions to add more exchange backends.
Created an exchange_backends folder with:
    * An abstract class exchange: all exchange wrappers should inherit from this abstract class, so that all implement the same minimal set of orders, at least.
    * An exchangeFactory class, that so far only works for the "kraken" exchange. More to be added.
    * A wrapKraken class, for the kraken exchange
    * A unit test for the exchange_backends submodule, only doing few tests for the kraken exchange.
Added a shell script run_unit_tests.sh, that runs the unittest discovery on all modules (so, folders with a __init__.py file)
Moved the former test.py file to simulator.py, to avoid messing with unittest discovery (the tools searches for files starting with "test").